[Feat] 플레이어와 악기 합치기
게임내에서 플레이어와 악기가 겹치는 부분이 많아
플레이어를 만드신 분과
악기를 만드신 분과 함께 리펙토링 해가며
둘의 스크립트를 합치는 작업을 하였다.
