+++ Assets/Settings/Renderer2D.asset
@@ -40,7 +40,7 @@ MonoBehaviour:
     blendMode: 0
   m_UseDepthStencilBuffer: 1
   m_UseCameraSortingLayersTexture: 1
-  m_CameraSortingLayersTextureBound: 0
+  m_CameraSortingLayersTextureBound: 783316077
   m_CameraSortingLayerDownsamplingMethod: 0
   m_MaxLightRenderTextureCount: 16
   m_MaxShadowRenderTextureCount: 1
