+++ Assets/Student/HSD/Scripts/Demo.cs
@@ -1,0 +1,24 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Demo : MonoBehaviour
+{
+    [SerializeField] private PlayerWeaponController controller;
+    [SerializeField] private MusicWeapon[] weapons;
+
+    private void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.Z))
+            controller.AddMusicWeapon(weapons[0]);
+
+        if (Input.GetKeyDown(KeyCode.X))
+            controller.AddMusicWeapon(weapons[1]);
+
+        if (Input.GetKeyDown(KeyCode.C))
+            controller.AddMusicWeapon(weapons[2]);
+
+        if (Input.GetKeyDown(KeyCode.V))
+            controller.AddMusicWeapon(weapons[3]);
+    }
+}
