+++ Assets/Student/HSD/Scripts/Player/PlayerInput.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerInput : MonoBehaviour
+{
+    private void Update()
+    {
+        
+    }
+}
