+++ Assets/Student/HSD/Scripts/Player/PlayerInput.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6bad77de0420bd6489a02a5e0fee0195
+guid: 5156e1db543d0bf41bed7dbbf816e735
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
