+++ Assets/Student/HSD/Scripts/Player/PlayerWeaponController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e4d0dc30dfa3dbd4cae40f4e56795f39
+guid: 196d307ab4b3c0746a5fed05de9255e5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
