+++ Assets/Student/HSD/Shader/ShockWave.mat
@@ -56,7 +56,7 @@ Material:
     m_Floats:
     - _ShockWaveStrength: -0.1
     - _Size: 0.05
-    - _WaveDistanceFromCenter: -0.1
+    - _WaveDistanceFromCenter: 0.351
     - _xSizeRatio: 0
     m_Colors:
     - _RingSpawnPosition: {r: 0.5, g: 0.5, b: 0, a: 0}
