+++ Assets/Student/JHT/Prefabs/Cymbals/CymbalsProjectile.prefab
@@ -71,8 +71,8 @@ SpriteRenderer:
   m_AutoUVMaxDistance: 0.5
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingLayer: 0
+  m_SortingLayerID: 305960425
+  m_SortingLayer: 2
   m_SortingOrder: 0
   m_Sprite: {fileID: 2747690628134850419, guid: b670ab75dde984907b8570040daa08c5, type: 3}
   m_Color: {r: 1, g: 0, b: 0.65316486, a: 1}
@@ -100,4 +100,6 @@ MonoBehaviour:
   maxTime: 0
   speed: 3
   IsPass: 0
+  rigid: {fileID: 0}
+  cymbalsSprite: {fileID: 0}
   LivingTime: 2
