+++ Assets/Student/JHT/Prefabs/Cymbals/ShaderCymbalsProjectile.prefab
@@ -75,7 +75,7 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: 2747690628134850419, guid: d0eebedf2e910144d8e3de13d9e2b5e1, type: 3}
+  m_Sprite: {fileID: 0}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
@@ -83,7 +83,7 @@ SpriteRenderer:
   m_Size: {x: 1, y: 0.8828125}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
-  m_WasSpriteAssigned: 1
+  m_WasSpriteAssigned: 0
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
 --- !u!114 &1447947333167568233
@@ -103,7 +103,7 @@ MonoBehaviour:
   IsPass: 0
   rigid: {fileID: -6474956957834104150}
   cymbalsSprite: {fileID: 917614916908520802}
-  LivingTime: 1
+  LivingTime: 6
 --- !u!50 &-6474956957834104150
 Rigidbody2D:
   serializedVersion: 4
