+++ Assets/Student/JHT/Prefabs/Trumpet/Trumpet.prefab
@@ -131,8 +131,8 @@ SpriteRenderer:
   m_AutoUVMaxDistance: 0.5
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: 0
-  m_SortingLayer: 0
+  m_SortingLayerID: -1801031537
+  m_SortingLayer: 3
   m_SortingOrder: 0
   m_Sprite: {fileID: 21300000, guid: 666b737749017194ea5fc01a3868ac57, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
