+++ Assets/Student/JHT/Sciptable/MusicWeapon/Cymbals.asset
@@ -28,5 +28,5 @@ MonoBehaviour:
   - {fileID: 499010178548250081, guid: 08a665cb9e369f641ae02f3c757e2cd8, type: 3}
   - {fileID: 8472133573239423094, guid: cb235c927c88f4546bde0acd5bca8298, type: 3}
   - {fileID: 5470499818708082747, guid: d3defe44ccca3814c8937d173f9677a4, type: 3}
-  Projectile: {fileID: -6229726346729069643, guid: 6a69481fd90741d41ad34f5ebf239d3b, type: 3}
+  Projectile: {fileID: 1447947333167568233, guid: 5e8c6730a581b4e49bfdc248774aed1c, type: 3}
   Projectiles: []
