+++ Assets/Student/JHT/Scripts/Weapon/Gun.cs
@@ -6,9 +6,8 @@ public class Gun : MusicWeapon
 {
     public override void Attack(Vector2 mousePosition)
     {
-        Vector2 pos = mousePosition.normalized;
-        Projectile obj = Instantiate(WeaponData.Projectile);
+        Vector2 pos = (mousePosition - (Vector2)transform.position).normalized;
+        Projectile obj = Instantiate(WeaponData.Projectile, transform.position, Quaternion.identity);
         obj.Init(pos);
     }
-
 }
